Refactors the signup_team_complete.jsx file into multiple files for each component, rather than have all components put in one file
